pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "social-engineer-arena"
version = "0.1.0"
description = "OpenEnv red/blue-team social engineering arena for RL post-training."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openenv-core>=0.1.0",
  "fastapi>=0.104.0",
  "uvicorn>=0.24.0",
  "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "matplotlib>=3.8.0"]
train = ["trl>=0.18.0", "transformers>=4.44.0", "datasets>=2.20.0", "accelerate>=0.33.0"]

[project.scripts]
server = "social_engineer_arena.server.app:main"

[tool.setuptools.packages.find]
include = ["social_engineer_arena*"]

[tool.setuptools.package-data]
social_engineer_arena = ["data/*.json"]
